Answer task 6

+++ 0x00-hello_world/6-size.c
@@ -9,8 +9,8 @@ int main(void)
 
 	printf("size of char: %d byte(s)\n", sizeof(char));
 	printf("size of int: %d byte(s)\n", sizeof(int));
-	printf("size of long int: %u byte(s)\n", sizeof(long long int));
-	printf("size of long long int: %u byte(s)\n", sizeof(long long int));
+	printf("size of long int: %zu byte(s)\n", sizeof(long int));
+	printf("size of long long int: %zu byte(s)\n", sizeof(long long int));
 	printf("size of float: %d byte(s)\n", sizeof(float));
 	return (0);
 }
